Update berlese-funnel-capacity.ttl
Got the definition from OpenAI.

Q: in ecology, what is berlese funnel capacity?
A: A Berlese funnel is a tool used in ecology to collect and analyze small invertebrates, such as mites and springtails, from soil or leaf litter. It is named after the Italian entomologist and botanist, Antonio Berlese, who developed the technique in the early 1900s.

The Berlese funnel consists of a funnel with a mesh screen at the bottom, which is placed over a container. The sample of soil or leaf litter is placed in the top of the funnel, and a heat source, such as a light bulb or heat lamp, is placed beneath the funnel to create a warm, dry environment. As the sample dries out, the small invertebrates will move towards the warmth and fall through the mesh screen into the container below, where they can be collected and studied.

The capacity of a Berlese funnel refers to the amount of sample that can be processed at one time. This will depend on the size of the funnel and the container, as well as the volume of the sample being collected. In general, a Berlese funnel has a capacity of a few hundred milliliters or less.