Update expected-number-of-trap-nights.ttl
Got the definition from OpenAI.

Q:in ecology, what is expected number of trap nights

A:The expected number of trap nights is a measure of the amount of effort that is put into a trapping study. It is calculated by multiplying the number of traps that are set out in a particular area by the number of nights that the traps are left out. For example, if 10 traps are set out in a particular area and they are left out for 5 nights, the expected number of trap nights would be 50. The expected number of trap nights is used to compare the effort that has been put into different trapping studies, or to compare the results of a trapping study to previous research on the same or similar species. It is important to consider the expected number of trap nights when interpreting the results of a trapping study, as the number of animals captured can be influenced by the amount of effort that has been put into the study.