Update aspect.ttl
Suggest splitting the concept 'aspect' into two separate concepts. Aspect and aspect ratio.

'Aspect' has the definition 'Identifies the downslope direction (orientation) of the slope. the reading is taken looking down the slope. Due north is recorded as 360 degrees. Aspect can only have a value of 1 to 360 degrees a value of 0 indicates that there is no slope.' Refers to landform and direction

'Aspect ratio' definition 'The proportional relationship between an image's width and height.' Refers to settings in a camera